(Partially) redo some files and add documentation.
- Remove some more unneeded code.
- Use VS documentation to document TwitchAPI, TwitchChatClient and TwitchWhisperClient functions and events.
- Improve parsing of certain objects to prevent exceptions [partially UNTESTED, please take your time to review].
- Improve TwitchAPI functions by not reusing as much code [partially UNTESTED, please take your time to review].
- Improve parameter naming in TwitchAPI (username -> channel where applicable).
- Enum integer binding is cool and reduces ifs, switches etc.

Again, this commit is partially UNTESTED and could break certain things. Please take your time to review this commit and make sure everything works. Also, the events got more fitting names so some things will be broken for some existing users. Make sure to notify your userbase of this.

Finally, this should have cleaned up the project quite a bit and put it more in-line with other projects. Hope it helps again. Sorry if this needs to be tested a lot, but I do not have access to certain things to test, including subscriptions etc.