The beginning of something nice.
This is the first commit in a series of commits that will ultimately expand and refactor the API portion of the project. This project isn't finished by a long shot, so this is not material to be merged as of writing, this is buggy as all hell most likely, although some stuff might work. All old API stuff hasn't changed at all so that's all still functional. I plan on adding EVERY SINGLE API ENDPOINT to this PR, which should be really darn useful. I will be working on this over multiple days, so yeah, be on the lookout for it. This commit is just to show you what I'm doing and to get feedback on certain endpoints' documentation etc. As always, hope it helps!

A list of changes so far:
- Move model files into Models/ folder (without changing namespace)
- Update all properties and make them all auto-accessors.
- Add TwitchFeedPost model. (WIP, not tested yet)
- Update TwitchStream model to add email and stream_key.
- Get rid of TwitchAPI.cs completely, we're doing something cooler here.
- Implement APIBase for helper functions and definitions.
- New Block API methods: GetBlockList, BlockUser, UnblockUser
- New Channel-Feed API methods: GetFeedPosts, AddFeedPost, GetFeedPost, DeleteFeedPost, ReactToFeedPost, DeleteFeedPostReaction
- New Channels API method: GetOwnTwitchChannel (including email and stream_key)
- Generate new API class stubs.
- Cleanup usings.