Added MessageEmote and MessageEmoteCollection
These lay the groundwork for doing emote replacement in messages.
As many users will not be concerned with emotes (bots, simple IRC
clients), these are fully standalone classes. Additional changes need to
be made to ChatMessage (and, I think, WhisperMessage) in order to add this
support - but those changes will be fully optional, with default
parameters ensuring backwards compatibility and ease-of-use.