Next Update
This update includes:-
Two base Secret Handlers - 1) Static and 2) Rotated these will be built on at a later date and I will include a filebased SecretHandler and an EF based SecretHandler
Authentication
Extension Methods for creating and using Authentication and ExtensionManagers in dotnet core.
Changes to API and other steps.
Added claims (this will be discussed in the docs.