Restructure of code
I didn't like the API and Extension classes where they were and it was an over complication. In this newer model it's still down to the dev however we have given a base abstract to work from.

I think I can improve the dotnat core DI a little but that can be worked on in the future

Started docs